requirements.txt
requests
tabulate
rich
python-dateutil
beautifulsoup4
html2text
lxml
curl_cffi
google-genai
